# Implement is_squarefree for QQ/ZZPolyRingElem