Cleaning out old branches; I made this last November, not sure
why I never submitted. The new methods replace generic code,
so this is purely about making things a bit faster.

Before:

    julia> S, y = polynomial_ring(QQ, :y); g = 2*(y+1)^2;

    julia> @btime is_squarefree(g)
      516.927 ns (10 allocations: 400 bytes)
    false

    julia> R, x = polynomial_ring(ZZ, :x); f = 4*(x^2+1);

    julia> @btime is_squarefree(f)
      1.329 μs (38 allocations: 1.82 KiB)
    false

After:

    julia> S, y = polynomial_ring(QQ, :y); g = 2*(y+1)^2;

    julia> @btime is_squarefree(g)
      53.596 ns (0 allocations: 0 bytes)
    false

    julia> R, x = polynomial_ring(ZZ, :x); f = 4*(x^2+1);

    julia> @btime is_squarefree(f)
      209.386 ns (4 allocations: 64 bytes)
    false