@fingolfin 2024-04-29
Aha. I guess the reason I did not submit it back then is that it runs into issue #1311 (which I completely forgot about), at least for `ZZPolyRingElem` (I am really glad we have these problematic cases covered by our test suite)