@fingolfin 2024-11-08
Fixes the `ZZPolyRingElem` case.